fix(log-cleanup): Let find handle globbing
Airflow instances with a lot of DAG runs create lots of small log files.
The `find` command is able to search for patterns, iterating over and
deleting them. However, the old way would use the shell's globbing
feature and pass the directories to check for as individual arguments.
This may caused an "Argument list too long" error.